+++ src/main/java/com/example/servlets/User/ThanhToanServlet.java
@@ -24,20 +24,19 @@ protected void doGet(HttpServletRequest request, HttpServletResponse response)
     HttpSession session = request.getSession();
     String user = (String) session.getAttribute("user");
     System.out.println("user: " + user + " đang xem trang thanh toán");
-
     List<CartItem> cart = (List<CartItem>) session.getAttribute("cart");
     request.setAttribute("cart", cart);
 
-    // Ép kiểu và kiểm tra tổng tiền
-    Double tongTienObj = (Double) session.getAttribute("tongTien");
-    double tongTien = (tongTienObj != null) ? tongTienObj : 0;
+
+    Integer tongTienObj = (Integer) session.getAttribute("tongTien");
+    int tongTien = (tongTienObj != null) ? tongTienObj : 0;
 
     if (tongTien <= 0) {
         request.setAttribute("thongBao", "Giỏ hàng của bạn đang trống hoặc có lỗi. Vui lòng thêm sản phẩm trước khi thanh toán.");
         request.getRequestDispatcher("cart.jsp").forward(request, response);
         return;
     }
-    String userValue = (String) session.getAttribute("user"); // Có thể là username hoặc email
+    String userValue = (String) session.getAttribute("user"); 
     if (userValue != null) {
         UserDAO userDAO = new UserDAO();
         User userInfo = userDAO.getUserByUsernameOrEmail(userValue);
@@ -45,14 +44,11 @@ protected void doGet(HttpServletRequest request, HttpServletResponse response)
         session.setAttribute("userInfo", userInfo);
 }
 
-
-    // Nếu hợp lệ, chuyển đến trang thanh toán
     request.getRequestDispatcher("thanhToan.jsp").forward(request, response);
 }
 
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        // Xử lý khi có yêu cầu POST (nếu cần)
         doGet(request, response);
     }
